Issue #7 ParamAndReturn types now create asserts
Resolving the issue that ParamAndReturn mocks don't have a asserts for InvokedWith.
The reason for the omission is that if something returns, you should assert against the
returned value.
This works well in most cases. There are some instances where a class is using the result into another another collaborator. In this case we're unable to check the specific return value.
As we're dealing with both collaborators, we're not getting a final result back which
will be dependent on the input. We can only verify that via a InvokedWith check. Unfortunate, but a needed update to how this library gets used.