#5 Updating processing to follow interface implementations.
Interfaces can implement other interfaces. Sometimes as placeholders, sometimes to hav ea common interface for a behavior. This wasn't support and was a pain to get into place. Now it's supported. The plugin will walk the interface inheritance and implement all of them.

it kinda borked the layout of the result I'd worked hard to achieve, but I'll file a bug for that and get to it later.